For edit form mode, ditch the top bar and just use the sidebar
